Parse the base directory from MSBuild to support finding errors in dependency projects.
Without loading the base directory of the error, next-error will
either display an invalid file with the same relative path from the
current directory of the user or query the user for the correct
path. This change makes visiting errors in dependency MSBuild projects
as easy as visiting errors in the current project.