Support compilation output from msbuild /maxcpucount (/m).
Lines of output from concurrent msbuild builds are, by default,
prefixed with something like "1>", where "1" is the number of the
msbuild node that caused the output to appear. This must be skipped
for the filename to be properly recognized.

This expands the regexp matching tests to check that the discovered
relative path is, indeed, the expected relative path. It also expands
the tests to account for all of the error/warning messages, including
the repeated messages summary.