Remove lots of compile-time evaluation.
This is a "take 1" on a csharp-mode with as few (eval-when-compile)
and (eval-and-compile) forms as possible.

Not only is it a source of confusion, but it also introduces
"soft" dependencies which cannot be resolved by the byte-compilation
system and this causes compilation warnings.

All in all, the changes should be fairly unintrusive, and we still have
an issue with our fairly icky (c-lang-defconst), but all the tests pass
and things still seem to work as they should.

This commit adresses the remaining warnings found in:
https://github.com/josteink/csharp-mode/issues/79