Move bobp checking in one level
It is always safe to use 'c-backward-syntactic-ws', so we can do that before
checking for bobp.  The reason this is smart is that the 'backward-char' can
fail with a 'beginning-of-buffer' error.  This causes indentation with 'C-x h
TAB' sometimes to fail.  Since a lambda function is super unlikely to be at
beginning of buffer, or only after some doc comments (syntactic ws), this is an
ok move.