Add more nodes to beginning/end-of-defun
I did not yet extract that list of declarations into a customizable variable,
but I think that is a reasonable thing to do at some point.

Not completely sure how this should behave.  Now tree-sitter defines things as a
tree (obviously...), so we need to be able to jump to siblings as well.
Add more nodes to beginning/end-of-defun

I did not yet extract that list of declarations into a customizable variable,
but I think that is a reasonable thing to do at some point.

Not completely sure how this should behave.  Now tree-sitter defines things as a
tree (obviously...), so we need to be able to jump to siblings as well.