Add support for multiline strings (#251)
* Add support for multiline strings

There is now a new type of support for multiline strings in CC Mode, and this is
a first attempt to support this.  The regexes are supplied from Alan Mackenzie,
the maintainer of CC Mode.  There are probably improvements to be made here, but
it seems to fix some slowness with regards to string handling.  How does it
work, you ask?

It is left as an exercise to the reader.