Implement patch from Stefan Monnier <monnier@iro.umontreal.ca>
Fix issues with (back)quoted lambda in csharp-mode.el:

--

Hi,

I was annoyed at the backquoted lambda in `csharp-mode.el` and tracked
down its origin to a bug in CC-mode.  While waiting for CC-mode to be
fixed, the patch below replaces this workaround with another one that
allows the code to be properly byte-compiled.